Move baseline into common module
