+++ common/src/main/java/org/corfudb/common/metrics/Gauge.java
@@ -1,4 +1,4 @@
-package org.corfudb.baseline.metrics;
+package org.corfudb.common.metrics;
 
 public interface Gauge<T extends Number> {
     T getValue();
